feat: api 호출 줄이기 위한 debounce 적용
