+++ src/hooks/useKeywordSuggestion.js
@@ -1,11 +1,13 @@
-import { useState, useEffect, useCallback } from 'react';
+import { useState, useEffect, useCallback, useRef } from 'react';
 import { apiClient } from '../apis/apiClient';
+import { getDataFromCache, setCache } from '../utils/cacheControl';
+import { debounce } from '../utils/debounce';
 
 const EXPIRE_TIME = 300; //300 seconds
 
 const useKeywordSuggestion = (keyword) => {
   const [suggestions, setSuggestions] = useState([]);
-  const fetchSuggestions = useCallback(async () => {
+  const fetchSuggestions = useCallback(async (keyword) => {
     try {
       if (keyword) {
         let cachedData = getDataFromCache(keyword);
@@ -23,32 +25,28 @@ const useKeywordSuggestion = (keyword) => {
     } catch (error) {
       console.error(error);
     }
-  }, [keyword]);
+  }, []);
+
+  const debounceRef = useRef(
+    debounce((keyword) => {
+      fetchSuggestions(keyword);
+    }, 500)
+  );
 
   useEffect(() => {
-    fetchSuggestions();
-  }, [fetchSuggestions]);
+    if (keyword) {
+      let cachedData = getDataFromCache(keyword);
+      if (cachedData) {
+        setSuggestions(cachedData);
+      } else {
+        debounceRef.current(keyword);
+      }
+    } else {
+      setSuggestions([]);
+    }
+  }, [keyword, debounceRef]);
 
   return [suggestions];
 };
-
-function getDataFromCache(keyword) {
-  const cache = localStorage.getItem(`suggestion-${keyword}`);
-  if (cache) {
-    const { data, expireTime: cacheExpireTime } = JSON.parse(cache);
-    if (new Date().getTime() < cacheExpireTime) {
-      return data;
-    }
-  }
-  return null;
-}
-
-function setCache(keyword, data, expireTime) {
-  const cache = {
-    data,
-    expireTime: new Date().getTime() + expireTime * 1000,
-  };
-  localStorage.setItem(`suggestion-${keyword}`, JSON.stringify(cache));
-}
 
 export default useKeywordSuggestion;
